+++ Source/Core/Actors/Events/EventGroup.cs
